species selection css refactor, replace step 4 and 5
todo: change px to rem, species selection image responsive
