+++ src/pages/Calculator/Steps/MixRatios/index.js
@@ -19,9 +19,8 @@ import {
   calculateAllValuesNECCC,
 } from "./../../../../shared/utils/calculate";
 import { updateSteps } from "../../../../features/stepSlice";
-import "./../steps.scss";
-import SeedsSelectedList from "../../../../components/SeedsSelectedList";
 import MixRatioSteps from "./form";
+import "./../steps.scss";
 
 const MixRatio = ({ council }) => {
   // themes
