+++ src/pages/Calculator/Steps/SeedingMethod/index.js
@@ -1,19 +1,22 @@
 //////////////////////////////////////////////////////////
-//                     Imports                          //
+//                    Imports                           //
 //////////////////////////////////////////////////////////
 
 import * as React from "react";
 import Grid from "@mui/material/Grid";
-import { useState, useEffect } from "react";
 import { useSelector, useDispatch } from "react-redux";
 import { Typography, Box } from "@mui/material";
+import Accordion from "@mui/material/Accordion";
+import AccordionSummary from "@mui/material/AccordionSummary";
+import AccordionDetails from "@mui/material/AccordionDetails";
+import ExpandMoreIcon from "@mui/icons-material/ExpandMore";
 import { useTheme } from "@mui/material/styles";
 import useMediaQuery from "@mui/material/useMediaQuery";
 
-import { VerticalSlider } from "../../../../components/VerticalSlider";
-import "./../steps.scss";
 import { updateSteps } from "../../../../features/stepSlice";
-import SeedsSelectedList from "../../../../components/SeedsSelectedList";
+import { seedingMethods } from "../../../../shared/data/dropdown";
+import { Dropdown } from "../../../../components/Dropdown";
+import "./../steps.scss";
 
 const SeedingMethod = ({ council }) => {
   // themes
@@ -25,15 +28,8 @@ const SeedingMethod = ({ council }) => {
   // useSelector for crops reducer data
   const dispatch = useDispatch();
   const data = useSelector((state) => state.steps.value);
-  const seedingMethod = data.seedingMethod;
-  const speciesSelection = data.speciesSelection;
+  const { seedingMethod, speciesSelection } = data;
   const seedsSelected = speciesSelection.seedsSelected;
-
-  const [dataLoaded, toggleDataLoaded] = useState(false);
-  const [min, setMin] = useState(0);
-  const [max, setMax] = useState(0);
-  const [seedingRateCoefficient, setSeedingRateCoefficient] = useState(0);
-  const [seedingRateAverage, setSeedingRateAverage] = useState(0);
 
   //////////////////////////////////////////////////////////
   //                      Redux                           //
@@ -49,132 +45,103 @@ const SeedingMethod = ({ council }) => {
   };
 
   //////////////////////////////////////////////////////////
-  //                    State Logic                       //
+  //                   State Logic                        //
   //////////////////////////////////////////////////////////
 
-  const updateSeedingRateAverage = async () => {
-    let average = 0;
-    seedsSelected.map((s, i) => {
-      average += Math.round(s.mixSeedingRate);
-    });
-    const minimum = average - average / 2;
-    const maximum = average + average / 2;
-    const coefficient =
-      average + (seedingMethod.managementImpactOnMix - 0.5) * average;
-    setMin(minimum);
-    setMax(maximum);
-    setSeedingRateAverage(average);
-    setSeedingRateCoefficient(
-      average + (seedingMethod.managementImpactOnMix - 0.5) * average
-    );
-    toggleDataLoaded(true);
+  const handleSeedingMethod = (e) => {
+    handleUpdateSteps("type", e.target.value);
   };
-  const updateManagementImpactOnMix = (e) => {
-    const percentage = e.target.value / seedingRateAverage - 0.5;
-    setSeedingRateCoefficient(e.target.value);
-    handleUpdateSteps("managementImpactOnMix", percentage);
-  };
-  const clearChart = () => {
-    toggleDataLoaded(false);
-  };
-  const renderSeedsSelected = () => {
-    return <SeedsSelectedList list={speciesSelection.seedsSelected} />;
-  };
-  const renderVerticalSlider = (seedingRate, handleUpdateSteps) => {
-    const marks = [];
 
-    marks.push({ value: min, label: `${min}` });
-    marks.push({ value: seedingRateCoefficient, label: `${seedingRate}` });
-    marks.push({ value: max, label: `${max}` });
-
+  const renderRightAccordian = (type, val) => {
     return (
-      <Grid xs={3} container justifyContent="center" alignItems="center">
-        <VerticalSlider
-          marks={marks}
-          value={seedingRate}
-          handleChange={(e) => {
-            updateManagementImpactOnMix(e);
-          }}
-        />
+      <Grid item xs={6} className="mix-seeding-rate-grid-right">
+        {council === "NECCC" && type !== "precision" ? (
+          <Typography>Not Recommended</Typography>
+        ) : (
+          <>
+            <Box
+              sx={{
+                width: "50px",
+                height: "50px",
+                padding: "11px",
+                margin: "0 auto",
+                backgroundColor: "#E5E7D5",
+                border: "#C7C7C7 solid 1px",
+                borderRadius: "50%",
+              }}
+            >
+              <Typography>{val}</Typography>
+            </Box>
+            <Typography>Lbs / Acre</Typography>
+          </>
+        )}
       </Grid>
     );
   };
 
-  //////////////////////////////////////////////////////////
-  //                    useEffect                         //
-  //////////////////////////////////////////////////////////
+  const renderAccordian = (seed) => {
+    return (
+      <Accordion xs={12} className="accordian-container">
+        <AccordionSummary
+          xs={12}
+          expandIcon={<ExpandMoreIcon />}
+          className="accordian-summary"
+        >
+          <Typography>{seed.label}</Typography>
+        </AccordionSummary>
+        <AccordionDetails className="accordian-details">
+          <Grid xs={12} container>
+            <Grid item xs={6} className="mix-seeding-rate-grid-left">
+              <Typography>Precision: </Typography>
+            </Grid>
+            {renderRightAccordian("precision", seed.precision)}
+            <Grid item xs={6} className="mix-seeding-rate-grid-left">
+              <Typography>Drilled: </Typography>
+            </Grid>
+            {renderRightAccordian("drilled", 1)}
+            <Grid item xs={6} className="mix-seeding-rate-grid-left">
+              <Typography>Broadcast(with Light Incorporation): </Typography>
+            </Grid>
+            {renderRightAccordian("broadcast", seed.broadcast)}
+            <Grid item xs={6} className="mix-seeding-rate-grid-left">
+              <Typography>
+                Aerial(or broadcast with no Light Incorporation{" "}
+                <span className="red-text">Not Recommended</span>):{" "}
+              </Typography>
+            </Grid>
+            {renderRightAccordian("aerial", seed.aerial)}
+          </Grid>
+        </AccordionDetails>
+      </Accordion>
+    );
+  };
 
-  useEffect(() => {
-    updateSeedingRateAverage();
-    return () => {
-      dataLoaded && clearChart();
-    };
-  }, []);
-
   //////////////////////////////////////////////////////////
-  //                      Render                          //
+  //                    Render                            //
   //////////////////////////////////////////////////////////
 
   return (
     <Grid container>
       <Grid item xs={12}>
-        <Typography variant="h2">Adjust Seeding Rate of Mix</Typography>
+        <Typography variant="h2">Select Seeding Method</Typography>
       </Grid>
-      <Grid
-        sx={{ paddingTop: 5 }}
-        container
-        xs={12}
-        justifyContent="center"
-        alignItems="center"
-      >
-        <Grid container xs={4} justifyContent="flex-end" alignItems="flex-end">
-          <Grid item xs={12}>
-            <Typography>
-              <Box
-                className="seeding-method-detail-container"
-                sx={{ marginBottom: "40px" }}
-              >
-                <Typography>Factors that lower Seeding Rate:</Typography>
-                <Typography>- Cost Saving</Typography>
-                <Typography>- Low Biomass</Typography>
-                <Typography>- Planting Green</Typography>
-              </Box>
-            </Typography>
-          </Grid>
-          <Grid item xs={12}>
-            <Box
-              className="seeding-method-label-container"
-              sx={{ marginBottom: "40px" }}
-            >
-              <Typography>Calculated Mix Seeding Rate</Typography>
-            </Box>
-          </Grid>
-          <Grid item xs={12}>
-            <Box className="seeding-method-detail-container">
-              <Typography>Factors that may raise Seeding Rate: </Typography>
-              <Typography>- Erosion Control</Typography>
-              <Typography>- Weed Supression</Typography>
-              <Typography>- Grazing</Typography>
-            </Box>
-          </Grid>
-        </Grid>
-        {dataLoaded &&
-          renderVerticalSlider(seedingMethod.mixSeedingRate, handleUpdateSteps)}
-        <Grid container xs={5} sx={{ height: "400px" }}>
-          <Grid item xs={12} justifyContent="flex-start">
-            <Box
-              className="seeding-method-label-container"
-              sx={{ marginBottom: "270px" }}
-            >
-              <Typography>High limit of Mix Seeding Rate</Typography>
-            </Box>
-          </Grid>
-          <Grid item xs={12} justifyContent="flex-end">
-            <Box className="seeding-method-label-container">
-              <Typography>Low limit of Mix Seeding Rate</Typography>
-            </Box>
-          </Grid>
-        </Grid>
+      <Grid item xs={12} padding={"15px"} className="">
+        <Dropdown
+          value={seedingMethod.type}
+          label={"Seeding Method: "}
+          handleChange={handleSeedingMethod}
+          size={12}
+          items={seedingMethods}
+        />
+      </Grid>
+      <Grid item xs={12}>
+        {seedsSelected.map((s, i) => {
+          return (
+            <Grid xs={12}>
+              <Grid item>{renderAccordian(s)}</Grid>
+            </Grid>
+          );
+        })}
       </Grid>
     </Grid>
   );
