+++ src/shared/data/dropdown.js
@@ -11,8 +11,8 @@ export const calculatorList = [
   "Site Conditions",
   "Species Selection",
   "Mix Ratios",
-  "Mix Seeding Rate",
   "Seeding Method",
+  "Mix Seeding Rate",
   "Seed Tag Info",
   "Review Mix",
   "Confirm Plan",
